requirements.txt
torch
opencv-python
Django
djangorestframework
boto3
botocore
cryptography
fernet